[actions] Check for Polkit authorization if the action requires so.
As to authenticate the user is_authorized() requires the DBus sender and connection objects,
the target action method is now cloned (by create_dbus_fixed_method()) and the sender and connection
keywords that DBus expects are inserted at run-time.

This is totally transparent and happens inside the decorator. The authorization check is also handled there.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>